move utils specific to test_submit_job.py into that file
