+++ tests/test_api/conftest.py
@@ -5,7 +5,6 @@ import pytest
 import responses
 
 from hyp3_api import CMR_URL, app, auth
-from hyp3_api.util import get_granules
 
 
 AUTH_COOKIE = 'asf-urs'
@@ -13,7 +12,6 @@ COSTS_URI = '/costs'
 JOBS_URI = '/jobs'
 USER_URI = '/user'
 
-DEFAULT_JOB_ID = 'myJobId'
 DEFAULT_USERNAME = 'test_username'
 DEFAULT_ACCESS_TOKEN = 'test_access_token'
 
@@ -32,29 +30,6 @@ def login(client, username=DEFAULT_USERNAME, access_token=DEFAULT_ACCESS_TOKEN):
         key=AUTH_COOKIE,
         value=auth.get_mock_jwt_cookie(username, lifetime_in_seconds=10_000, access_token=access_token),
     )
-
-
-def make_job(granules=None, name='someName', job_type='RTC_GAMMA', parameters=None):
-    if granules is None:
-        granules = ['S1B_IW_SLC__1SDV_20200604T082207_20200604T082234_021881_029874_5E38']
-    if parameters is None:
-        parameters = {}
-    job = {'job_type': job_type, 'job_parameters': {'granules': granules, **parameters}}
-    if name is not None:
-        job['name'] = name
-
-    return job
-
-
-def submit_batch(client, batch=None, validate_only=None):
-    if batch is None:
-        batch = [make_job()]
-    payload = {
-        'jobs': batch,
-    }
-    if validate_only is not None:
-        payload['validate_only'] = validate_only
-    return client.post(JOBS_URI, json=payload)
 
 
 def make_db_record(
@@ -95,7 +70,7 @@ def make_db_record(
     return record
 
 
-def setup_requests_mock_with_given_polygons(granule_polygon_pairs):
+def setup_mock_cmr_response_for_polygons(granule_polygon_pairs):
     cmr_response = {
         'feed': {
             'entry': [
@@ -104,19 +79,3 @@ def setup_requests_mock_with_given_polygons(granule_polygon_pairs):
         }
     }
     responses.add(responses.POST, CMR_URL_RE, json.dumps(cmr_response))
-
-
-def setup_requests_mock(batch):
-    granule_polygon_pairs = [
-        (
-            granule,
-            [
-                [
-                    '3.871941 -157.47052 62.278873 -156.62677 62.712959 -151.784653 '
-                    '64.318275 -152.353271 63.871941 -157.47052'
-                ]
-            ],
-        )
-        for granule in get_granules(batch)
-    ]
-    setup_requests_mock_with_given_polygons(granule_polygon_pairs)
